Update README.md
Added a requirements section to specify that this library requires a PHP version of 5.4.0 or above because of the use of the short array syntax (e.g. $post = []; on line 28 of RedirectController.php). This is important to note because Nexcess, a popular Magento host, still sets up servers with PHP 5.3.24 by default at this time.